Welcome Tooltips, now in python!
This patch include python bindings to elm tooltip support.

We created one python object Widget_Item, that the python elm items must
inherit. In this, we moved the items of genlist, list and toolbar to
inherit from Widget_Item and included them the tooltips methods.

The test of tooltip have the same comportament of tooltip test in
elementary_test. So, is very easy compare the minor diferences between
the languages to tooltips.

Author:    Fabiano Fidencio <fidencio@profusion.mobi>
           and me.

git-svn-id: svn+ssh://svn.enlightenment.org/var/svn/e/trunk@52252 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33