annoyances--: proper zoom changes, add auto-fit.
after actually using it to browse a directory of pictures I realized
auto-fit is something we need, and need it by default.

as we auto-fit, the zoom range makes no sense anymore... and zoom
should be percentual changes, not absolute. So now zoom increments and
decrements of 20% of the previous value, and there is no lower/upper
limits anymore.



git-svn-id: svn+ssh://svn.enlightenment.org/var/svn/e/trunk@54057 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33