Remove elm_notepad
Notepad had a big issue: users couldn't access the entry, so notepad
was something near to an useless widget.

We needed to choose between wrap all entry functions or add these file
save / load and autosave features to entries.

As discussed on the emails list, we decided to go with the second
option, so we won't have an extra widget with lots of wrap functions.

Notepad was internally just an elm_entry with a scroll, so using
a scrolled entry with a file set will do the job.

I've added a test called Entry Notepad that looks like the removed
Notepad test. But we will see the improvements - you can clear the entry
or whatever you want, because the entry is the object you are creating.

If you were using elm_notepad before, just s/notepad/scrolled_entry
and almost everything should work fine. Another change was the function
file_save(obj, file, format), is now just file_save(obj), and the
file and format used are defined by file_set().

At last, some getters were added.



git-svn-id: svn+ssh://svn.enlightenment.org/var/svn/e/trunk@54145 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33