Cleanup icon up handling
It seems that originally the "selected" event wasn't to be fired on dirs
during single click. This makes it impossible for fwin to open a new
window when single clicking a dir, so some ugly hacks has been added.
"The right fix (tm)" is to always emit "selected", even on dirs, and let
the listener decide what to do with the file.

git-svn-id: svn+ssh://svn.enlightenment.org/var/svn/e/trunk@54229 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33