[enjoy] Improve cover art management/caching
This creates the "covers" table inside "media.db", to cache the path to
each album's cover art. With this, we significantly reduce the amount of
time required to open a previously-seen album screen. This has been
implemented primarily to support downloading cover art from LastFM
(and not hammering their servers or the user's bandwidth).

Currently, LastFM support is enabled by default and we're not complying
with their terms of service (we need to include a badge on our UI if
we're using their data -- that's why we have a "origin" column in the
"covers" database). The badge stuff will be fixed soon, and enabling/
disabling LastFM support will happen whenever we have some settings
dialog working.

git-svn-id: svn+ssh://svn.enlightenment.org/var/svn/e/trunk@54605 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33