From: hyoyoung.chang@samsung.com
As we talked last week, I made a version of adding external image in
elm_entry. It seeks a local file if prefix is 'file://' then it loads image and adds it. If image file can't be loaded, adding
wft emoticon.

patch in with minor fix for path and an added actual test case in elm
entry test.




git-svn-id: svn+ssh://svn.enlightenment.org/var/svn/e/trunk@55134 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33