From: 김지훈 <imfine98@gmail.com>
elm_photocam widget does not support keyboard input.
This patch makes elm_photocam be able to process keyboard input such
as Left,
Right, Up, Down, PgDn, and PgUp key.
In addition, image will be zoomed in or out when '+' or '-' key in
keypad is pressed.

Please review it and let me know if there is any problem in this patch.



git-svn-id: svn+ssh://svn.enlightenment.org/var/svn/e/trunk@55842 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33