add fps debug display to comp...
to toggle interactively:
  killall -USR2 enlightenment 
or
  ctrl+alt+shift+F
or just change it in the config panel - the longer the average
interval is, the more accurate it will be, but the less often you get
updates. you can choose the screen corner here too. :) also note that
if comp doesnt get a stream of input (frames to render) then it's also
not going to tel you much interesting about its "best possible
framerate" as its likely got lots of spare time to do more, but it
just never got any to do. so this is useful if:

1. there is a continual strem of input (video playback, animated gui,
continual user input).
2. the longer the rolling average framecount is, the more accurate it
will be.



git-svn-id: svn+ssh://svn.enlightenment.org/var/svn/e/trunk@56260 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33