From: Leif Middelschulte <leif.middelschulte@gmail.com> Subject: [PATCH] website
Website patch.

Currently we use a bad search string with google as it searches the
entire website instead of just the docs domain as the search page
suggests. If people like it would be easy to integrate the search
results into e.g. a dynamic iframe on the docs page so users can stay
on e.org while searching the docs.

BR,

Leif


git-svn-id: svn+ssh://svn.enlightenment.org/var/svn/e/trunk@57945 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33