Evas textblock: Speed up rendering. Find the start paragraph for all.
Until now all the format/text drawing loops went through all the
paragraphs skipping all the ones that are outside the clipping/object
zone. This changeset find the first paragraph to be renedred first so we
don't walk all the paragraphs for nothing. This speeds up rendering of
the ends of very big textblocks in a very noticable manner.

git-svn-id: svn+ssh://svn.enlightenment.org/var/svn/e/trunk@59404 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33