Tue Mar 15 23:03:57 2011                                           mej
Fix for CVE-2011-0409 (CERT VU#285156), a use-after-free error in the
XIM code.  This only affects versions where XIM support is compiled in
(which it is by default).  There are no known exploits for this bug,
but it is theoretically exploitable.  Thanks to Jonathan Brossard and
the team at Toucan System for responsibly disclosing this
vulnerability and to CERT for assisting with coordination and
disclosure.
----------------------------------------------------------------------


git-svn-id: svn+ssh://svn.enlightenment.org/var/svn/e/trunk@59413 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33