Tue Mar 15 23:08:26 2011                        Michael Jennings (mej)
Fix for CVE-2011-0768, an off-by-one error in handling large pixmap
filenames which resulted in an overflow of a single NUL character if
the filename exceeded PATH_MAX bytes.  This bug is NOT exploitable.
Again, thanks to Jonathan Brossard and the team at Toucan System for
responsibly disclosing this vulnerability and to CERT for assisting
with coordination and disclosure.
----------------------------------------------------------------------


git-svn-id: svn+ssh://svn.enlightenment.org/var/svn/e/trunk@59414 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33