From: Jihoon Kim <jihoon48.kim@samsung.com> Subject: [E-devel] [PATCH] Add some APIs for supporting virtual keyboard in Ecore_IMF
For supporting virtual keyboard, I'd like to add some APIs. The detail
description of each API is included in the patch file as doxygen format.

In the attached patch, the reason why we add the subprefix 'input_panel_'
related to virtual keyboard is that input method can be soft keyboard or voice
input or image captured by camera.



git-svn-id: svn+ssh://svn.enlightenment.org/var/svn/e/trunk@59894 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33