ecore/doc - Move some examples to examples.dox
Also remove references to examples that don't exist anymore.

More examples are going to be explained, and removing them from Ecore.h
will improve the readability of that file. This is the same that was
done to Elementary.

I'm going to move all the examples reference to this file, and should
have them being pointed by functions that use them too.



git-svn-id: svn+ssh://svn.enlightenment.org/var/svn/e/trunk@60598 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33