evas/evas_object_image - modified to keep the uv consistency with normal evas_object.
Although the uv range of normal object is be localized to the it's geometry, 

but the uv range of the image object is be localized it's image size but not geometry size. 

In that case, user should find out the objects type and set the uv value differently between normal and image

why they do that?



git-svn-id: svn+ssh://svn.enlightenment.org/var/svn/e/trunk@61577 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33