THEMES/detorious: rework genlist theme.
Remove cruft (odd is now a signal, not a different group), refactor
the common part and simplify the theme.

I still don't like it much, rather go with a list like the one in
"everything" (e17), but until we settle on a look and feel, let's use
this.

Enjoy is now 50% better :-)



git-svn-id: svn+ssh://svn.enlightenment.org/var/svn/e/trunk@62766 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33