Evas textblock: Fixed an issue when there are no available fonts.
The problem was that our rbtree implementation doesn't support chaining
duplicates and with no fonts, all the paragraphs had the same y coord (0).
Changing the paragraph index to work on line_no for sorting fixes it,
as line numbers are also correct when there are no fonts.

Thanks a lot to KaKaRoTo for letting me know about this issue.

git-svn-id: svn+ssh://svn.enlightenment.org/var/svn/e/trunk@63188 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33