Elementary map : a bug fix for zoom-out in elm_map by Kim Yunhan
I wrote a patch that handles a bug while zooming-out in elm_map.
When I try to zoom out, some tiles are broken.
But it is hard to notice because broken frame disappears quickly.

I investigated in a few days.
And I realize that there are something wrong.
When map is zoomed out, a tile is shrunk by evas_object_resize().
But evas_map handles its texture by just its origin image size not a shrunk size.
If evas_object's width & height is shrunk, I have to handle for its texture.


git-svn-id: svn+ssh://svn.enlightenment.org/var/svn/e/trunk@63509 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33