evas: Use clists to store the render recalculation list
Rather than trying to avoid removing the list element that is
currently being processed, keep two lists and move elements
to the processed list before recalculating them.

Remove items from the list head only, and always append them
to the tail.

Use the fact that an item can be removed from a clist without
needing to know which list it is in.

Signed-off-by: Mike McCormack <mj.mccormack@samsung.com>

git-svn-id: svn+ssh://svn.enlightenment.org/var/svn/e/trunk@64030 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33