From: Jaehwan Kim <jaehwan.kim.neo@gmail.com>
At first, try to execute elementary_test and drag down the list for bounce.
Then as soon as release the mouse, click the list continuously and fast.
The list will not be clicked because it cannot be clicked during the bounce
animation. The bounce animation time is fixed and it is reset again, when
the mouse is down and up. In the result, if we click it continuously, we can
not choose the list item. (During animation, list can not be clicked).
I changed the function "_smart_bounce_x_animator" and "_smart_bounce_
y_animator". I fixed the bounce time will be changed by remaining distance.



git-svn-id: svn+ssh://svn.enlightenment.org/var/svn/e/trunk@64407 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33