From: Doyoun Kang <doyoun.kang@samsung.com> Subject: [E-devel] [Patch] e_config config save patch
I found a problem in e_config. There is no code to check file size when the e makes the
configuration backup files. Even if the file size is 0, the e copies it to backup file.
So, I added code to check the size of the file. Please check the attached patch file.



git-svn-id: svn+ssh://svn.enlightenment.org/var/svn/e/trunk@64684 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33