Evas textblock: Fixed a bug with textblock ellipsis.
There was a possible segfault because we don't check if the current item
is a text item or a format item. I just removed the loop which triggered it
because it's not needed anyway, and now it works. Removing the loop also
let me remove some code that was only needed in the case of a loop.

git-svn-id: svn+ssh://svn.enlightenment.org/var/svn/e/trunk@64816 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33