e_dbus/connman: prepare for upcoming release, review API.
Let's prepare for E_Connman.h release with E_DBus.

Lucas De Marchi already sync'ed the API with recent upstream service,
removing APIs no longer exist, adding new ones.

I've noticed that some lists were using Eina_List, some const char **,
some would allocate and some not. This commit will settle on
"const char **" with no allocations on getters, same for setters.

Also, in order to make it clear that the API is dependent on upstream
service API we're requiring the users to acknowledge the unstability with:
{{{
   #define E_CONNMAN_I_KNOW_THIS_API_IS_SUBJECT_TO_CHANGE 1
}}}

We'll submit a patch to connman so it exports the API version, then we
can check it during runtime as well.



git-svn-id: svn+ssh://svn.enlightenment.org/var/svn/e/trunk@65192 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33