Ecore: fix Shift, Control and Alt keys detection
Shift (left/right), Control and Alt are now detected.
AltGr is not, as it's actually a Control (left) key,
immediately followed by an Alt key, and I don't know how
to see know if the left Control kye is the one pressed by
the user or the one thrown by AltGr.

modifiers are not set too. Will be set soon

git-svn-id: svn+ssh://svn.enlightenment.org/var/svn/e/trunk@65415 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33