From: cnook <kimcinoo@gmail.com> Subject: [E-devel] [Patch] elc_ctxpopup, add restacking feature
I have attached two files, one is the patch(*.diff) for adding
restacking feature to the ctxpopup.
The other is test example(*.TEST_EXAMPLE). You can test on the
elementary_test > ctxpopup > "Ctxpopup with user content"
And click the button which is labeled "Enlightenment". Then label will
overlap the ctxpopup.. The ctxpopup will raise in a little time.
Please review the patch and if there is no problem, please throw the
test example away! :-)




git-svn-id: svn+ssh://svn.enlightenment.org/var/svn/e/trunk@66415 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33