Added EvasGL support for software backend finally.
* This feature requires libOSMesa to be installed.   

One caveat with OSMesa is that a surface config (ie. Depth Buffer and etc)
is associated with a context rather than a surface, which is the case 
in EvasGL.  So for now, when a user specifies a surface config, it gets 
associated with the first context that the surface does a make current to.
For typical usage case, this shouldn't be a prolem. Will need to fix it 
eventually.




git-svn-id: svn+ssh://svn.enlightenment.org/var/svn/e/trunk@66896 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33