python-evas: fix buildbot warnings.
yay! buildbot spotted a nice error showing unused functions, it was
not using them because it was using the wrong functions (canvas
instead of object). At the end the IMAGE_UNLOADED is also an object
method, not a canvas one.

also fix some stupid warning at evas_object_image_rotate.c, gcc is
stupid enough to not notice src_alpha is just used if has_alpha is
defined (I guess llvm will catch this "fix" as a warning!)



git-svn-id: svn+ssh://svn.enlightenment.org/var/svn/e/trunk@67344 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33