welcome eina_model.
This is a simple to use generic hierarchical data access. It includes
properties, children, reference counting and notifications (callbacks).

It can be understood in the same lines as GObject or PyObject, but I
believe the best usage is just to provide data models, not generic
Object Oriented Programming in C.

It misses most of documentation, although the important bits are
there. Need help here :-/



git-svn-id: svn+ssh://svn.enlightenment.org/var/svn/e/trunk@67405 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33