expedite: remove software_sdl backend and use buffer instead.
This result in an overall drop of 17% in performance. I think it
may be due to some stuff related to HWSURFACE, the fact it use
double buffer or that maybe SDL_FillRect was faster for cleaning up
memory. If someone is really interested by this performance drop,
better fix them with buffer engine.


git-svn-id: svn+ssh://svn.enlightenment.org/var/svn/e/trunk@68351 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33