elev8: Allow loading modules in runtime
There are now two kinds of modules: native (.so) and JavaScript (.js). Both
modules can be loaded using the same interface and is completely transparent
to the application developer, by using the global require() function.  For
example, to load the elementary module:

    elm = require('elm');

This will try to open 'libelm.so' in the current directory, then 'libelm.so'
in elev8's module directory.  If that fails, it will try to open 'elm.js' in
the current directory, then 'elm.js' in elev8's directory.  Should any of
these attempts work, it will save the module in an internal cache and return
a reference to an object containing all exported symbols from that module.
The module loading path can be changed by changing the MODPATH environment
variable in the same way one would set the PATH environment variable.

To create a native module, one should create a
RegisterModule(Handle<ObjectTemplate>target) function, and manually add
exported symbols to the target object. Be sure to ``export "C"'' it or else
dlsym() will fail.

JavaScript modules works pretty much the same way: they're created and
executed in a separate context so that they do not mess up with the rest of
the application (for instance, it is very easy to forget a ``var'' keyword,
making a variable global -- imagine the mess a for (i = 0; i < 10; i++)
derp() would make).  A global (to the module's context) object called
``exports'' must be used to export functions and globals:

    exports.pi = 3.14159265;
    exports.calculate_the_answer = function() {
	return Math.ceil((1337 * exports.pi) / 100.0);
    }

Elev8 does not load all modules automatically anymore -- this is now
controlled by the init.js script, which is loaded and interpreted at
startup.  This can be used to load other modules, define useful constants,
this sort of thing.  Care should be taken when using this file, though, since
it is compiled and evaluated in the global context.

Also, support for Eina_Module has been temporarily removed; it will be back,
don't worry about it.  This means that module shutdown functions have been
temporarily removed as well.  There are more important cleanups to be made
in Elev8, so this will be postponed until further notice.



git-svn-id: svn+ssh://svn.enlightenment.org/var/svn/e/trunk@68583 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33