From: Davide Andreoli <dave@gurumeditation.it> Subject: [E-devel] Emotion buffer size patch
Hi, here is a patch for emotion, it add a new function to retrive the
status of the buffer while playing online stream.
It is implemented only for the gstreamer backend, the xine one
do not play at all here. What about the generic one?

(xine does play - i implameneted this with the xine module, and
generic is given a func that always returns 1.0 for now).



git-svn-id: svn+ssh://svn.enlightenment.org/var/svn/e/trunk@68684 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33