From: cnook <kimcinoo@gmail.com> Subject: [E-devel] [Patch][elementary] elc_ctxpopup using elm_list internally.
The patch is for using elm_list as a item controller of elc_ctxpopup.
I have learned that there was an opinion to change the internal
structure
of ctxpopup using the list.
It would be able to match up to your expectation. but it is just a draft
version of using list in ctxpopup.
So, It would be pretty good to me if you give any feedback on this.
Thanks.




git-svn-id: svn+ssh://svn.enlightenment.org/var/svn/e/trunk@69561 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33