From: cnook <kimcinoo@gmail.com> Subject: [E-devel] [patch][elementary] elc_popup, restack issue
The elc_popup does not raise to the top of its layer even though
evas_object_raise() is called with the popup object.
Because the elm_widget_resize_object_set(); makes notify object which is
internally used in the popup have a smart parent.
Ultimately, the callback function for EVAS_CALLBACK_RESTACK of the
notify
is not called.
So I removed elm_widget_resize_object_set(); from the popup and added
some
implementations to show the popup properly.
Moreover, I have added test code also in the attached patch. Then,
please
review the patch and give any feedbacks. Thanks.



git-svn-id: svn+ssh://svn.enlightenment.org/var/svn/e/trunk@69790 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33