elev8: Put HandleScope in wrapper of getters and setters Always make sure getters and setters cleanup their stuff by putting a HandleScope in wrapper. This way getters and setters do not need to define one and we make sure we are not leaking variables to the next scope.
101 of HandleScope:

When you create Local<...> variables, it uses the storage defined by the
last scope. When the destructor of the scope runs, it marks all objects
handled by it as "can be freed" so GC is allowed to take them back when
it wants. HandleScope can be only allocated on stack and when the C++
scope of this variable finishes, the destructor runs (in C++ parlance,
it uses RAII).

If you don't use HandleScopes, the Local<...> variable will be allocated
on whenever scope was declared previously. For getters and setters we
always know that: they are in the static wrappers.

Patch by: Lucas De Marchi <lucas.demarchi@profusion.mobi>



git-svn-id: svn+ssh://svn.enlightenment.org/var/svn/e/trunk@70032 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33