ecore_thread: add example of code using ecore_thread.
fractale-vector: use gcc vector extention to do its math.
fractale-fp: use eina_f16p16 to do its math.
fractale-orc: this is currently not working as expected. Things need
to be fixed on orc side so that I can came back to this example and
fix it. So no need to bother with it at this point.


git-svn-id: svn+ssh://svn.enlightenment.org/var/svn/e/trunk@70561 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33