[Elm] Creating base smart class, to be extended  properly by real widgets. I'm also creating a *compatibility* class on top of it, which extends the old (unique) one to its old components, so that everything goes unnoticed for the widgets.
Next, we'll be accessing class functions for widget hooks *on the base
class*, which will be proxied back to the instance ones.

Then, one by one the widgets will be translated to a proper
hierarchical smart class parentage, making extensibility and reuse
much better.

Last, but not least, since we'll be moving these hooks to the class,
where (almost) everything is a function, let's adapt to the future --
take away the "_func" suffix from them.

NB: there is ONE hook and associated data which was kept on the base,
for now: "show_region" stuff. I did that because scrollers need that
in order to work, right now. Keep in mind, though, the scrollers will
be reworked, so that fields ought to be changed soon, too.



git-svn-id: svn+ssh://svn.enlightenment.org/var/svn/e/trunk@70704 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33