evas: remove librsvg svg loader.
NOTE: as librsvg is a massive source of bugs in e17, it is now
removed from evas. You can still use librsvg by using the
evas_generic_loader. Please not that you need to properly delete
it from your disk if you don't use a package manager. The file to
remove :
/*/lib/evas/modules/loaders/svg/linux-gnu-i686-1.2.*/module.so



git-svn-id: svn+ssh://svn.enlightenment.org/var/svn/e/trunk@71223 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33